Implement log fetching in modal
The log fetching functionality has been updated in this commit. Instead of displaying logs on the same page, they are now fetched and displayed in a modal window. Additionally, the route for fetching logs was updated to include a 'type' parameter which defines the format in which logs should be returned. A 'logsmodal' HTML definition was created to handle displaying of logs in the frontend.